meteo/requirements.txt
asgiref==3.8.1
Django==5.2.1
pyserial==3.5
sqlparse==0.5.3
